requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
Django==5.1
django-ninja==1.3.0
itsdangerous==2.2.0
passlib[argon2]==1.7.4
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg-pool==3.2.2
pydantic-settings==2.3.1
PyJWT==2.8.0
tzdata==2024.1
python-decouple==3.8